# By tinkl (6) and others
# Via tinkl (6) and others
* 'appStore' of github.com:nicolastinkl/AI2020OS:
  update popToRootView Action...
  修复 搜索首页 ui细节
  解决提交订单闪退的问题
  UPdate
  update: 修复拍照问题
  修复 warning
  修改测试数据
  添加了字服务列表视图，修改了订单列表的点击效果
  update
  添加了图片形式的字服务卡片视图
  update
  修复searchLabel换行问题
  去掉警告
  修复label的size不对bug
  fix ui bugs

Conflicts:
	Veris/Business/CustomerCenter/View/ServiceExecute/AITimelineNowTableViewCell.swift
	Veris/Business/ProudctInfo/View/AIServerProviderView.swift

+++ Veris/Business/CustomerCenter/View/BuyerViews/AIHorizontalBarChart.m
@@ -133,7 +133,7 @@
     CGFloat y = (_maxHeight - fontSize) / 2;
     CGRect frame = CGRectMake(x, y, width, fontSize);
     
-    _numberLabel = [AIViews normalLabelWithFrame:frame text:[NSString stringWithFormat:@"%ld", _number] fontSize:fontSize color:Color_MiddleWhite];
+    _numberLabel = [AIViews normalLabelWithFrame:frame text:[NSString stringWithFormat:@"%ld", (long)_number] fontSize:fontSize color:Color_MiddleWhite];
     _numberLabel.lineBreakMode = NSLineBreakByTruncatingTail;
     [self addSubview:_numberLabel];
 }
